style: apply linter

+++ src/mpi-mode/row-compute.h
@@ -14,4 +14,4 @@
  *
  * @return Consumed time.
  */
-double mpi_process_by_rows(int rank, int size, const struct p_args *args, const struct filter_mix *filters);
+double mpi_process_by_rows(int8_t rank, int8_t size, const struct p_args *args, const struct filter_mix *filters);
